[김하연] 기능수정: 좋아요 이력 테이블 수정
