+++ models/index.js
@@ -5,8 +5,7 @@ const Board             = require('./boards');
 const Category          = require('./categories');
 const Directory         = require('./directories');
 const Image             = require('./image');
-const LikeRecordOfPost  = require('./likeRecordOfPost');
-const LikeRecordOfReply = require('./likeRecordOfReply');
+const LikeRecord        = require('./likeRecord');
 const Post              = require('./posts');
 const Reply             = require('./replies');
 const Report            = require('./reports');
@@ -28,8 +27,7 @@ db.Board             = Board;
 db.Category          = Category;
 db.Directory         = Directory;
 db.Image             = Image;
-db.LikeRecordOfPost  = LikeRecordOfPost;
-db.LikeRecordOfReply = LikeRecordOfReply;
+db.LikeRecord        = LikeRecord;
 db.Post              = Post;
 db.Reply             = Reply;
 db.Report            = Report;
