--- models/likeRecordOfPost.js
@@ -1,22 +1,0 @@
-const Sequelize = require('sequelize');
-
-module.exports = class LikeRecordOfPost extends Sequelize.Model{
-    static init(sequelize) {
-        return super.init({
-        }, {
-            sequelize,
-            timestamps: true,
-            underscored: true,
-            modelName: 'LikeRecordOfPost',
-            paranoid: false,
-            charset: 'utf8',
-            collate: 'utf8_general_ci',
-        });
-
-
-    }
-    static associate(db){
-        db.LikeRecordOfPost.belongsTo(db.Post, { onDelete: 'CASCADE', foreignKey:'postId', targetKey: 'id'});
-        db.LikeRecordOfPost.belongsTo(db.User,{ onDelete: 'CASCADE', foreignKey: 'userId', targetKey: 'id'});
-    }
-}
