+++ models/replies.js
@@ -29,5 +29,7 @@ module.exports = class Reply extends Sequelize.Model{
         db.Reply.belongsTo(db.Post, { onDelete: 'CASCADE', foreignKey: 'postId', targetKey: 'id' });
         db.Reply.belongsTo(db.Reply, { as: 'ReReply', foreignKey: 'parentId'});
         db.Reply.belongsTo(db.User, { foreignKey: 'userId', targetKey: 'id'});
+        // Post 연결 관계 설정할 때만 delete CASCADE가 적용이 안 되는 문제 발생
+        // mysql에서 수동으로 ON DELETE CASCADE 설정해 줌
     }
 }
